graphqlbackend: move external service exclusion logic to `backend` package.
Test plan:
Tests updated to reflect changes.
